WIP improvements

+++ src/Pents.ScreenHelper.Sharp/Models/ScreenCoordsDto.cs
@@ -1,3 +1,3 @@
 namespace Pents.ScreenHelper.Sharp.Models;
 
-public record ScreenCoordsDto(int X, int Y, float Confidence);
+public record ScreenCoordsDto(int X, int Y, double Confidence);
